updating price matching
In an effort to use true prices for cards and objects in MTGO collection. previously the program assumed there was 1 object that wasnt being found a price (event tickets) and just plugged in its value of 1 usd/tix. but actually the program also fails to match Foils to their appropriate prices.